🐰왜 이미지 바꾸면 썸머노트 날아가니

+++ Uglory_Admin/src/main/java/edu/kh/ugloryA/product/controller/ProductController.java
@@ -178,7 +178,15 @@ public class ProductController {
 	@ResponseBody
 	@GetMapping("/option/delete")
 	public int deleteOption(int optionCode) {
-		return service.deleteOption(optionCode);
+		
+		int result = 0;
+		int count = service.countOption(optionCode);
+		
+		if(count == 0) {
+			result = service.deleteOption(optionCode);
+		}
+		
+		return result;
 	}
